+++ src/modules/SX127x/SX1278.cpp
@@ -102,25 +102,25 @@ int16_t SX1278::setBandwidth(float bw) {
   uint8_t newBandwidth;
 
   // check allowed bandwidth values
-  if(fabs(bw - 7.8) <= 0.001) {
+  if(fabsf(bw - 7.8) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_7_80_KHZ;
-  } else if(fabs(bw - 10.4) <= 0.001) {
+  } else if(fabsf(bw - 10.4) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_10_40_KHZ;
-  } else if(fabs(bw - 15.6) <= 0.001) {
+  } else if(fabsf(bw - 15.6) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_15_60_KHZ;
-  } else if(fabs(bw - 20.8) <= 0.001) {
+  } else if(fabsf(bw - 20.8) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_20_80_KHZ;
-  } else if(fabs(bw - 31.25) <= 0.001) {
+  } else if(fabsf(bw - 31.25) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_31_25_KHZ;
-  } else if(fabs(bw - 41.7) <= 0.001) {
+  } else if(fabsf(bw - 41.7) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_41_70_KHZ;
-  } else if(fabs(bw - 62.5) <= 0.001) {
+  } else if(fabsf(bw - 62.5) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_62_50_KHZ;
-  } else if(fabs(bw - 125.0) <= 0.001) {
+  } else if(fabsf(bw - 125.0) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_125_00_KHZ;
-  } else if(fabs(bw - 250.0) <= 0.001) {
+  } else if(fabsf(bw - 250.0) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_250_00_KHZ;
-  } else if(fabs(bw - 500.0) <= 0.001) {
+  } else if(fabsf(bw - 500.0) <= 0.001) {
     newBandwidth = RADIOLIB_SX1278_BW_500_00_KHZ;
   } else {
     return(RADIOLIB_ERR_INVALID_BANDWIDTH);
@@ -623,7 +623,7 @@ void SX1278::errataFix(bool rx) {
   // sensitivity optimization for 500kHz bandwidth
   // see SX1276/77/78 Errata, section 2.1 for details
   Module* mod = this->getMod();
-  if(fabs(SX127x::bandwidth - 500.0) <= 0.001) {
+  if(fabsf(SX127x::bandwidth - 500.0) <= 0.001) {
     if((frequency >= 862.0) && (frequency <= 1020.0)) {
       mod->SPIwriteRegister(0x36, 0x02);
       mod->SPIwriteRegister(0x3a, 0x64);
@@ -639,49 +639,49 @@ void SX1278::errataFix(bool rx) {
   // figure out what we need to set
   uint8_t fixedRegs[3] = { 0x00, 0x00, 0x00 };
   float rxFreq = frequency;
-  if(fabs(SX127x::bandwidth - 7.8) <= 0.001) {
+  if(fabsf(SX127x::bandwidth - 7.8) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x48;
     fixedRegs[2] = 0x00;
     rxFreq += 0.00781;
-  } else if(fabs(SX127x::bandwidth - 10.4) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 10.4) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x44;
     fixedRegs[2] = 0x00;
     rxFreq += 0.01042;
-  } else if(fabs(SX127x::bandwidth - 15.6) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 15.6) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x44;
     fixedRegs[2] = 0x00;
     rxFreq += 0.01562;
-  } else if(fabs(SX127x::bandwidth - 20.8) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 20.8) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x44;
     fixedRegs[2] = 0x00;
     rxFreq += 0.02083;
-  } else if(fabs(SX127x::bandwidth - 31.25) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 31.25) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x44;
     fixedRegs[2] = 0x00;
     rxFreq += 0.03125;
-  } else if(fabs(SX127x::bandwidth - 41.7) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 41.7) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x44;
     fixedRegs[2] = 0x00;
     rxFreq += 0.04167;
-  } else if(fabs(SX127x::bandwidth - 62.5) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 62.5) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x40;
     fixedRegs[2] = 0x00;
-  } else if(fabs(SX127x::bandwidth - 125.0) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 125.0) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x40;
     fixedRegs[2] = 0x00;
-  } else if(fabs(SX127x::bandwidth - 250.0) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 250.0) <= 0.001) {
     fixedRegs[0] = 0b00000000;
     fixedRegs[1] = 0x40;
     fixedRegs[2] = 0x00;
-  } else if(fabs(SX127x::bandwidth - 500.0) <= 0.001) {
+  } else if(fabsf(SX127x::bandwidth - 500.0) <= 0.001) {
     fixedRegs[0] = 0b10000000;
     fixedRegs[1] = mod->SPIreadRegister(0x2F);
     fixedRegs[2] = mod->SPIreadRegister(0x30);
