+++ src/modules/SX128x/SX128x.cpp
@@ -734,13 +734,13 @@ int16_t SX128x::setBandwidth(float bw) {
     return(RADIOLIB_ERR_WRONG_MODEM);
   }
 
-  if(fabs(bw - 203.125) <= 0.001) {
+  if(fabsf(bw - 203.125) <= 0.001) {
     this->bandwidth = RADIOLIB_SX128X_LORA_BW_203_125;
-  } else if(fabs(bw - 406.25) <= 0.001) {
+  } else if(fabsf(bw - 406.25) <= 0.001) {
     this->bandwidth = RADIOLIB_SX128X_LORA_BW_406_25;
-  } else if(fabs(bw - 812.5) <= 0.001) {
+  } else if(fabsf(bw - 812.5) <= 0.001) {
     this->bandwidth = RADIOLIB_SX128X_LORA_BW_812_50;
-  } else if(fabs(bw - 1625.0) <= 0.001) {
+  } else if(fabsf(bw - 1625.0) <= 0.001) {
     this->bandwidth = RADIOLIB_SX128X_LORA_BW_1625_00;
   } else {
     return(RADIOLIB_ERR_INVALID_BANDWIDTH);
@@ -1349,7 +1349,7 @@ RadioLibTime_t SX128x::getTimeOnAir(size_t len) {
       uint32_t N_symbolPreamble = (this->preambleLengthLoRa & 0x0F) * (uint32_t(1) << ((this->preambleLengthLoRa & 0xF0) >> 4));
 
       // calculate the number of symbols
-      N_symbol = (float)N_symbolPreamble + coeff1 + 8.0 + ceil(RADIOLIB_MAX((int16_t)(8 * len + N_bitCRC - coeff2 + N_symbolHeader), (int16_t)0) / (float)coeff3) * (float)(this->codingRateLoRa + 4);
+      N_symbol = (float)N_symbolPreamble + coeff1 + 8.0 + ceilf((float)RADIOLIB_MAX((int16_t)(8 * len + N_bitCRC - coeff2 + N_symbolHeader), (int16_t)0) / (float)coeff3) * (float)(this->codingRateLoRa + 4);
 
     } else {
       // long interleaving - abandon hope all ye who enter here
